+++ pydymenu/__init__.py
@@ -1,8 +1,7 @@
 from pydymenu.rofi import rofi
-from pydymenu.system import has_bin
 from pydymenu.fzf import fzf
 
 __author__ = "Mikey Garcia, @gikeymarcia"
 __version__ = "0.2.0"
 
-__all__ = ["fzf", "has_bin", "rofi"]
+__all__ = ["fzf", "rofi"]