+++ pydymenu/fzf.py
@@ -15,6 +15,7 @@ def fzf(
     prompt: Union[str, None] = None,
     multi: bool = False,
     case_sensitive: bool = False,
+    preview: Union[str, None] = None,
 ) -> Union[list[str], None]:
     """Take list of strings and returns selection list or None."""
     if not has_bin("fzf"):
@@ -24,6 +25,7 @@ def fzf(
         "prompt": prompt,
         "multi": multi,
         "case_sensitive": case_sensitive,
+        "preview": preview,
     }
     return _run_fzf_with_list(items, **_fzf_dict)
 
@@ -35,29 +37,30 @@ def _run_fzf_with_list(menu_entries: list[str], **kwargs) -> Union[None, list[st
         input=newline_joined_bytestream(menu_entries),
         stdout=sp.PIPE,
     )
-    console.log(f"{_fzf_process = }")
+    # console.log(f"{_fzf_process = }")
     return process_stdout(_fzf_process)
 
 
 def process_fzf_opts(options_dict: dict) -> list[str]:
     """Takes a dictionary of fzf options and returns the command line flags."""
-    # print(f"opts: {options_dict}")
-    fzf_flags = []
-    if prompt := options_dict.get("prompt", None):
-        fzf_flags.extend(["--prompt", prompt])
+
+    def get(key: str):
+        """Return a dict value from options_dict or None if missing."""
+        return options_dict.get(key, None)
+
+    flags = []
+    if prompt := get("prompt"):
+        flags.extend(["--prompt", prompt])
     # mutli-select
-    multi = options_dict.get("multi", None)
-    multi_mode = "--multi" if multi else "--no-multi"
-    fzf_flags.append(multi_mode)
-    # case sensitive (default False)
-    if options_dict["case_sensitive"]:
-        fzf_flags.append("+i")
+    multi = "--multi" if get("multi") else "--no-multi"
+    flags.append(multi)
+    if get("case_sensitive"):
+        flags.append("+i")
     else:
-        fzf_flags.append("-i")
-    # TODO: preview
-    # TODO: print_query
-    # print(f"adds: {fzf_flags}")
-    return fzf_flags
+        flags.append("-i")
+    if prev := get("preview"):
+        flags.extend(["--preview", prev])
+    return flags
 
 
 # vim: foldlevel=5: